+++ features/helper/apps/action/base.go
@@ -1,4 +1,4 @@
-package iosaction
+package action
 
 import (
 	"github.com/golang-automation/features/helper"
