+++ features/helper/apps/action/driver.go
@@ -1,22 +1,15 @@
-package androiddriver
+package action
 
 import (
 	"github.com/golang-automation/features/helper"
-	android "github.com/golang-automation/features/helper/apps/android"
 	"github.com/sclevine/agouti/appium"
 )
-
-/*AndroidPage : base page apk*/
-type AndroidPage struct {
-	Page   android.AndroidDriver
-	Device *appium.Device
-}
 
 /*Device global variable*/
 var Device *appium.Device
 
 /*StartDriver : start android driver*/
-func (s *AndroidPage) StartDriver() error {
+func (s *AppPage) StartDriver() error {
 	err := s.Page.Driver.Start()
 	helper.LogPanicln(err)
 
@@ -24,7 +17,7 @@ func (s *AndroidPage) StartDriver() error {
 }
 
 /*NewDevice : create new android device*/
-func (s *AndroidPage) NewDevice() *appium.Device {
+func (s *AppPage) NewDevice() *appium.Device {
 	var err error
 
 	Device, err = s.Page.Driver.NewDevice()
