+++ features/helper/apps/action/send_keys.go
@@ -1,4 +1,4 @@
-package iosaction
+package action
 
 import (
 	"github.com/golang-automation/features/helper"
