--- features/helper/apps/android/action/page.go
@@ -1,8 +1,0 @@
-package androidaction
-
-import android "github.com/golang-automation/features/helper/apps/android/driver"
-
-/*Page : page actions*/
-type Page struct {
-	Action android.AndroidPage
-}
