--- features/helper/apps/ios/action/page.go
@@ -1,8 +1,0 @@
-package iosaction
-
-import ios "github.com/golang-automation/features/helper/apps/ios/driver"
-
-/*Page : page actions*/
-type Page struct {
-	Action ios.IOSPage
-}
