--- features/helper/apps/ios/driver/driver.go
@@ -1,34 +1,0 @@
-package iosdriver
-
-import (
-	"github.com/golang-automation/features/helper"
-	ios "github.com/golang-automation/features/helper/apps/ios"
-	"github.com/sclevine/agouti/appium"
-)
-
-/*IOSPage : base page app*/
-type IOSPage struct {
-	Page   ios.IOSDriver
-	Device *appium.Device
-}
-
-/*Device global variable*/
-var Device *appium.Device
-
-/*StartDriver : start ios driver*/
-func (s *IOSPage) StartDriver() error {
-	err := s.Page.Driver.Start()
-	helper.LogPanicln(err)
-
-	return nil
-}
-
-/*NewDevice : create new ios device*/
-func (s *IOSPage) NewDevice() *appium.Device {
-	var err error
-
-	Device, err = s.Page.Driver.NewDevice()
-	helper.LogPanicln(err)
-
-	return Device
-}
