+++ features/helper/apps/ios/ios_connect.go
@@ -6,11 +6,6 @@ import (
 	"github.com/sclevine/agouti/appium"
 )
 
-/*IOSDriver : iOS driver*/
-type IOSDriver struct {
-	Driver *appium.WebDriver
-}
-
 /*Driver global variable*/
 var Driver *appium.WebDriver
 var p *properties.Properties
